+++ morphocluster/cli.py
@@ -6,6 +6,7 @@ from getpass import getpass
 import click
 import flask_migrate
 import h5py
+import numpy as np
 import pandas as pd
 import tqdm
 from etaprogress.progress import ProgressBar
@@ -158,8 +159,8 @@ def init_app(app):
 
                 # TODO: Use UPDATE ... RETURNING to get the number of affected rows
 
-                bar = ProgressBar(len(object_ids), max_width=40)
-                obj_iter = iter(zip(object_ids, vectors))
+                progress = tqdm.tqdm(total=len(object_ids), unit="obj")
+                obj_iter = iter(zip(object_ids, vectors))  # type: ignore
                 while True:
                     chunk = tuple(itertools.islice(obj_iter, 1000))
                     if not chunk:
@@ -172,9 +173,8 @@ def init_app(app):
                         ],
                     )
 
-                    bar.numerator += len(chunk)
-                    print(bar, end="\r")
-                print()
+                    progress.update(len(chunk))
+                progress.close()
 
                 # TODO: In the end, print a summary of how many objects have a feature vector now.