+++ CHANGELOG.rst
@@ -2,3 +2,5 @@ Current (0.2.0pre)
 ==================
 
 - Optimize growing performance by performing distance calculations in PostgreSQL (#48)
+
+- Grow largest clusters first (#54)